requirements.txt
numpy==1.26.4
openai==1.33.0
python-dotenv==1.0.1
tiktoken==0.7.0
flask==3.0.3
opencv-python==4.10.0.82
uuid==1.30